Comprehensive decimal support for all numeric inputs
HTML Template fixes:
- Add step attributes to all number inputs that were missing them
- Set appropriate min/max ranges for year and rate inputs
- Add pattern attribute to frequency field to ensure decimal validation
- Prevent form submission on Enter key press in service modal

JavaScript fixes:
- Fix parseInt usage for projection years to parseFloat for decimal support
- Ensure all numeric parsing preserves decimal values

Key improvements:
- Frequency per year now properly accepts decimals like 1.5, 0.5, 0.33
- All year inputs have proper integer step (step="1")
- All rate/percentage inputs have decimal step (step="0.01" or step="0.1")
- Added form submission prevention to avoid premature form submission
- Consistent min/max ranges across all year inputs (2020-2100)

+++ static/app.js
@@ -1512,7 +1512,7 @@ function updateSettingsSection() {
 
 function saveApplicationSettings() {
     appSettings.defaultBaseYear = parseInt(document.getElementById('default-base-year').value);
-    appSettings.defaultProjectionYears = parseInt(document.getElementById('default-projection-years').value);
+    appSettings.defaultProjectionYears = parseFloat(document.getElementById('default-projection-years').value);
     appSettings.defaultDiscountRate = parseFloat(document.getElementById('default-discount-rate').value);
     appSettings.autoCalculate = document.getElementById('auto-calculate').checked;
     appSettings.showWelcome = document.getElementById('show-welcome').checked;
